requirements.txt
flask
pandas
numpy
openai
langchain
langchain-community
huggingface-hub
sentence-transformers
faiss-cpu
python-dotenv
fuzzywuzzy
langchain-huggingface